move document components to document folder
