+++ app/components/geoblacklight/document/citation_component.html.erb
