+++ app/components/geoblacklight/document/citation_component.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module Geoblacklight
+  module Document
+    class CitationComponent < Blacklight::Document::CitationComponent
+      attr_reader :document
+      with_collection_parameter :document
+
+      def initialize(document:)
+        @document = document
+        super
+      end
+
+      def citation
+        document.geoblacklight_citation(solr_document_url(document))
+      end
+    end
+  end
+end
