+++ app/components/geoblacklight/document/sidebar_component.html.erb
