+++ app/components/geoblacklight/document/sidebar_component.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Geoblacklight
+  module Document
+    class SidebarComponent < Blacklight::Document::SidebarComponent
+    end
+  end
+end
