--- app/components/geoblacklight/sidebar_component.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module Geoblacklight
-  class SidebarComponent < Blacklight::Document::SidebarComponent
-  end
-end
