+++ app/views/catalog/_citation.html.erb
@@ -8,7 +8,7 @@
       <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
     </div>
     <%= t('blacklight.tools.citation_info') %>
-    <% citation = render Geoblacklight::CitationComponent.with_collection(@documents) if @documents %>
+    <% citation = render Geoblacklight::Document::CitationComponent.with_collection(@documents) if @documents %>
     <div class="d-inline">
       <%= citation %>
     </div>
